Added another Input Select for filter row styling consistancy
