+++ src/EventLogExpert/Shared/Components/FilterRow.razor
@@ -3,20 +3,10 @@
     {
         <div class="filter-content">
             <span>Filter On: </span>
-            <select class="input-filter-dropdown" @bind="@Value.FilterType">
-                @foreach (FilterType type in Enum.GetValues(typeof(FilterType)))
-                {
-                    <option value="@type">@type.ToString()</option>
-                }
-            </select>
+            <FilterSelect @bind-Value="Value.FilterType" InputType="FilterType" />
 
             <span>Comparison: </span>
-            <select class="input-filter-dropdown" @bind="@Value.FilterComparison">
-                @foreach (FilterComparison type in Enum.GetValues(typeof(FilterComparison)))
-                {
-                    <option value="@type">@type.ToFullString()</option>
-                }
-            </select>
+            <FilterSelect @bind-Value="Value.FilterComparison" InputType="FilterComparison" />
 
             <span>Value: </span>
             <div>
@@ -26,7 +16,7 @@
                 }
                 else
                 {
-                    <FilterSelect @bind-Value="@Value.FilterValue" Type="Value.FilterType" />
+                    <FilterValueSelect @bind-Value="@Value.FilterValue" Type="Value.FilterType" />
                 }
             </div>
         </div>
