--- src/EventLogExpert/Shared/Components/FilterSelect.razor.css
@@ -1,42 +1,0 @@
-﻿.input-select {
-    position: relative;
-
-    & input { margin: 0; }
-
-    & i {
-        position: absolute;
-        right: 2px;
-        cursor: pointer;
-
-        color: #fff;
-
-        &[data-toggle="true"] { transform: translateY(10%) rotate(-180deg); }
-    }
-}
-
-.dropdown-list {
-    border: 1px solid;
-    background-color: var(--background-dark);
-    z-index: 99;
-    position: absolute;
-    top: 2rem;
-    width: 100%;
-    max-height: 240px;
-    overflow-y: auto;
-    visibility: hidden;
-    transition: visibility 0.1s linear;
-
-    &[data-toggle="true"] { visibility: visible; }
-
-    & > div {
-        text-align: left;
-        margin: 0;
-        padding: 0 0 0 .5rem;
-
-        color: #fff;
-
-        &:hover {
-            background-color: var(--background-darkgray);
-        }
-    }
-}
