+++ src/EventLogExpert/wwwroot/css/app.css
@@ -109,12 +109,34 @@ a,
     text-align-last: center;
 }
 
-    .input-filter-dropdown { width: 100px; }
+    .input-filter-dropdown {
+        width: 120px;
+        text-align: center;
+        cursor: default;
+        user-select: none;
+    }
 
     .input-filter-select,
     .input-filter-description {
         width: 250px;
         text-align: center;
+    }
+
+    .input-select {
+        position: relative;
+        margin: 0 .5rem;
+
+        & input { margin: 0; }
+
+        & i {
+            position: absolute;
+            right: 2px;
+            cursor: pointer;
+
+            color: #fff;
+
+            &[data-toggle="true"] { transform: translateY(10%) rotate(-180deg); }
+        }
     }
 
     .input-filter-datetime { width: 200px; }
@@ -188,6 +210,36 @@ div.table-divider {
     /* col-resize is currently bugged in WebView2 */
     /* https://github.com/dotnet/maui/issues/9960 */
     div.table-divider:hover { cursor: e-resize !important; }
+
+
+
+.dropdown-list {
+    user-select: none;
+    border: 1px solid;
+    background-color: var(--background-dark);
+    z-index: 99;
+    position: absolute;
+    top: 2rem;
+    width: 100%;
+    max-height: 240px;
+    overflow-y: auto;
+    visibility: hidden;
+    transition: visibility 0.1s linear;
+
+    &[data-toggle="true"] { visibility: visible; }
+
+    & > div {
+        text-align: left;
+        margin: 0;
+        padding: 0 0 0 .5rem;
+
+        color: #fff;
+
+        &:hover {
+            background-color: var(--background-darkgray);
+        }
+    }
+}
 
 /*@supports (-webkit-touch-callout: none) {
     .status-bar-safe-area {
